+++ ktlint-ruleset-standard/src/main/kotlin/com/github/shyiko/ktlint/ruleset/standard/SpacingAroundCommaRule.kt
@@ -11,7 +11,7 @@ class SpacingAroundCommaRule : Rule("comma-spacing") {
 
     override fun visit(node: ASTNode, autoCorrect: Boolean,
             emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit) {
-        if (node is LeafPsiElement && (node.textMatches(",") || node.textMatches(";")) && !node.isPartOfString() &&
+        if (node is LeafPsiElement && node.textMatches(",") && !node.isPartOfString() &&
             PsiTreeUtil.nextLeaf(node) !is PsiWhiteSpace) {
             emit(node.startOffset + 1, "Missing spacing after \"${node.text}\"", true)
             if (autoCorrect) {