+++ CHANGELOG.md
@@ -36,6 +36,20 @@ respectively (`--ruleset-*` will be removed in 1.0.0).
 - `ktlint-intellij-idea-integration` binary   
 ([Intellij IDEA integration](https://github.com/shyiko/ktlint#option-1-recommended) task is now included in `ktlint` (as `ktlint --apply-to-idea`)).
 
+## [0.8.3] - 2017-06-19
+
+### Fixed
+
+- "Missing spacing after ";"" at the end of package declaration ([#59](https://github.com/shyiko/ktlint/issues/59)).
+- "Unused import" false positive (`setValue`) ([#55](https://github.com/shyiko/ktlint/issues/55)).
+- `get`/`set`ter spacing ([#56](https://github.com/shyiko/ktlint/pull/56)).
+
+## [0.8.2] - 2017-06-06
+
+### Fixed
+
+- "Unused import" false positive (`getValue`) ([#54](https://github.com/shyiko/ktlint/issues/54)).
+
 ## [0.8.1] - 2017-05-30
 
 ### Fixed
@@ -155,7 +169,9 @@ set in `[*{kt,kts}]` section).
 
 ## 0.1.0 - 2016-07-27
 
-[0.9.0]: https://github.com/shyiko/ktlint/compare/0.8.1...0.9.0
+[0.9.0]: https://github.com/shyiko/ktlint/compare/0.8.3...0.9.0
+[0.8.3]: https://github.com/shyiko/ktlint/compare/0.8.2...0.8.3
+[0.8.2]: https://github.com/shyiko/ktlint/compare/0.8.1...0.8.2
 [0.8.1]: https://github.com/shyiko/ktlint/compare/0.8.0...0.8.1
 [0.8.0]: https://github.com/shyiko/ktlint/compare/0.7.1...0.8.0
 [0.7.1]: https://github.com/shyiko/ktlint/compare/0.7.0...0.7.1